requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pep517==0.12.0
pip-tools==6.5.1
psycopg2-binary==2.9.3
six==1.16.0
tomli==2.0.1
Werkzeug==2.0.3